'Atomic' Packaging of Screens + Screen Renaming
