+++ app/src/main/java/elder/ly/mobile/ui/components/SpecialtyList.kt
@@ -7,7 +7,7 @@ import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.dp
-import elder.ly.mobile.ui.screens.SpecialtyItem
+import elder.ly.mobile.ui.screens.search.SpecialtyItem
 
 @Composable
 fun SpecialtyList(specialties: List<String>, onRemove: (String) -> Unit) {
