+++ app/src/main/java/elder/ly/mobile/ui/screens/personalinfo/InformacoesAdicionais.kt
@@ -1,6 +1,5 @@
-package elder.ly.mobile.ui.screens
+package elder.ly.mobile.ui.screens.personalinfo
 
-import android.icu.lang.UCharacter.LineBreak
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -18,7 +17,6 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
