+++ app/src/main/java/elder/ly/mobile/ui/screens/profile/Perfil.kt
@@ -1,4 +1,4 @@
-package elder.ly.mobile.ui.screens
+package elder.ly.mobile.ui.screens.profile
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.clickable
