+++ app/src/main/java/elder/ly/mobile/ui/screens/searchresult/PesquisaCuidador2.kt
@@ -1,4 +1,4 @@
-package elder.ly.mobile.ui.screens
+package elder.ly.mobile.ui.screens.searchresult
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
